`mapLifecycleListener` submodule extraction
